add rouserConfig 1laba
