+++ src/components/AppRoutes/AppRoutes.tsx
@@ -1,0 +1,7 @@
+import { useRoutes } from "react-router-dom";
+import routesConfig from "../../config/routesConfig";
+function AppRoutes() {
+  return useRoutes(routesConfig);
+}
+
+export default AppRoutes;
